Reduce complexity of platform.ts
